requirements.txt
fastapi
uvicorn
prisma
asyncpg
qdrant-client
pydantic-settings
pydantic[email]
python-dotenv
python-jose[cryptography]
langchain-community
pypdf
docx2txt
langchain-text-splitters
langchain-core
langchain-openai
langgraph
langgraph-checkpoint-postgres
psycopg[binary,pool]
httpx
tiktoken
docx2txt
sqlglot
python-docx
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pytest
pytest-asyncio
pymysql
